requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
agents>=0.1.0 